requirements.txt
alabaster==0.7.12
amqp==2.5.2
appdirs==1.4.3
argon2-cffi==19.1.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
billiard==3.6.3.0
black==19.3b0
celery==4.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.4
django-allauth==0.39.1
django-celery-beat==1.5.0
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.1
django-model-utils==3.2.0
django-redis==4.10.0
django-timezone-field==4.0
djangorestframework==3.10.2
docutils==0.16
entrypoints==0.3
factory-boy==2.12.0
Faker==4.0.2
flake8==3.7.8
flower==0.9.3
gunicorn==20.0.4
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
ipdb==0.12.2
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
kombu==4.6.8
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.720
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2==2.8.3
ptyprocess==0.6.0
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
pyparsing==2.4.6
pytest==5.1.0
pytest-django==3.5.1
pytest-sugar==0.9.2
python-crontab==2.4.1
python-dateutil==2.8.1
python-slugify==3.0.3
python3-openid==3.1.0
pytz==2019.2
redis==3.3.7
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
termcolor==1.1.0
text-unidecode==1.2
toml==0.10.0
tornado==5.1.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.9
Werkzeug==0.14.1
whitenoise==4.1.3
wrapt==1.11.2
zipp==3.1.0
